+++ RayTracerChallenge/src/Utils.h
@@ -25,7 +25,7 @@ namespace RayTracer
 		int barWidth = 70;
 
 		std::cout << "[";
-		int pos = barWidth * progress;
+		int pos = (int)(barWidth * progress);
 		for (int i = 0; i < barWidth; ++i) {
 			if (i < pos) std::cout << "=";
 			else if (i == pos) std::cout << ">";
